--- tests/Todo/Feature/Http/Controllers/Staff/CommandControllerTest.php
@@ -1,174 +1,0 @@
-<?php
-
-namespace Tests\Feature\Http\Controllers\Staff;
-
-use Tests\TestCase;
-
-/**
- * @see \App\Http\Controllers\Staff\CommandController
- */
-class CommandControllerTest extends TestCase
-{
-    /**
-     * @test
-     */
-    public function clear_all_cache_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get('dashboard/commands/clear-all-cache');
-
-        $response->assertRedirect(withInfo(trim(\Artisan::output())));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function clear_cache_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get('dashboard/commands/clear-cache');
-
-        $response->assertRedirect(withInfo(trim(\Artisan::output())));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function clear_config_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get('dashboard/commands/clear-config-cache');
-
-        $response->assertRedirect(withInfo(trim(\Artisan::output())));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function clear_route_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get('dashboard/commands/clear-route-cache');
-
-        $response->assertRedirect(withInfo(trim(\Artisan::output())));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function clear_view_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get('dashboard/commands/clear-view-cache');
-
-        $response->assertRedirect(withInfo(trim(\Artisan::output())));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function index_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.commands.index'));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.command.index');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function maintance_disable_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get('dashboard/commands/maintance-disable');
-
-        $response->assertRedirect(withInfo(trim(\Artisan::output())));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function maintance_enable_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get('dashboard/commands/maintance-enable');
-
-        $response->assertRedirect(withInfo(trim(\Artisan::output())));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function set_all_cache_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get('dashboard/commands/set-all-cache');
-
-        $response->assertRedirect(withInfo(trim(\Artisan::output())));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function test_email_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get('dashboard/commands/test-email');
-
-        $response->assertRedirect(withInfo(trim(\Artisan::output())));
-
-        // TODO: perform additional assertions
-    }
-
-    // test cases...
-}
