--- tests/Todo/Feature/Http/Controllers/Staff/InviteControllerTest.php
@@ -1,32 +1,0 @@
-<?php
-
-namespace Tests\Feature\Http\Controllers\Staff;
-
-use Tests\TestCase;
-
-/**
- * @see \App\Http\Controllers\Staff\InviteController
- */
-class InviteControllerTest extends TestCase
-{
-    /**
-     * @test
-     */
-    public function index_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.invites.index'));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.invite.index');
-        $response->assertViewHas('invites');
-        $response->assertViewHas('invitecount');
-
-        // TODO: perform additional assertions
-    }
-
-    // test cases...
-}
