--- tests/Todo/Feature/Http/Controllers/Staff/NoteControllerTest.php
@@ -1,71 +1,0 @@
-<?php
-
-namespace Tests\Feature\Http\Controllers\Staff;
-
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Tests\TestCase;
-
-/**
- * @see \App\Http\Controllers\Staff\NoteController
- */
-class NoteControllerTest extends TestCase
-{
-    use RefreshDatabase;
-
-    /**
-     * @test
-     */
-    public function destroy_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $note = factory(\App\Models\Note::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->delete(route('staff.notes.destroy', ['id' => $note->id]));
-
-        $response->assertRedirect(withSuccess('Note Has Successfully Been Deleted'));
-        $this->assertDeleted($staff);
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function index_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.notes.index'));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.note.index');
-        $response->assertViewHas('notes');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function store_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $note = factory(\App\Models\Note::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->post(route('staff.notes.store', ['username' => $note->username]), [
-            // TODO: send request data
-        ]);
-
-        $response->assertRedirect(withErrors($v->errors()));
-
-        // TODO: perform additional assertions
-    }
-
-    // test cases...
-}
