--- tests/Todo/Feature/Http/Controllers/Staff/TypeControllerTest.php
@@ -1,125 +1,0 @@
-<?php
-
-namespace Tests\Feature\Http\Controllers\Staff;
-
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Tests\TestCase;
-
-/**
- * @see \App\Http\Controllers\Staff\TypeController
- */
-class TypeControllerTest extends TestCase
-{
-    use RefreshDatabase;
-
-    /**
-     * @test
-     */
-    public function create_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.types.create'));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.type.create');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function destroy_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $type = factory(\App\Models\Type::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->delete(route('staff.types.destroy', ['id' => $type->id]));
-
-        $response->assertRedirect(withSuccess('Type Successfully Deleted'));
-        $this->assertDeleted($staff);
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function edit_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $type = factory(\App\Models\Type::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.types.edit', ['id' => $type->id]));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.type.edit');
-        $response->assertViewHas('type');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function index_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->get(route('staff.types.index'));
-
-        $response->assertOk();
-        $response->assertViewIs('Staff.type.index');
-        $response->assertViewHas('types');
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function store_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->post(route('staff.types.store'), [
-            // TODO: send request data
-        ]);
-
-        $response->assertRedirect(withErrors($v->errors()));
-
-        // TODO: perform additional assertions
-    }
-
-    /**
-     * @test
-     */
-    public function update_returns_an_ok_response()
-    {
-        $this->markTestIncomplete('This test case was generated by Shift. When you are ready, remove this line and complete this test case.');
-
-        $type = factory(\App\Models\Type::class)->create();
-        $user = factory(\App\Models\User::class)->create();
-
-        $response = $this->actingAs($user)->patch(route('staff.types.update', ['id' => $type->id]), [
-            // TODO: send request data
-        ]);
-
-        $response->assertRedirect(withErrors($v->errors()));
-
-        // TODO: perform additional assertions
-    }
-
-    // test cases...
-}
